Update Namespaces and directories
