+++ .idea/.idea.BaccaratTracker/.idea/.gitignore
