+++ .idea/.idea.BaccaratTracker/.idea/indexLayout.xml
