--- .idea/.idea.SamplePlugin/.idea/.name
@@ -1,1 +1,0 @@
-SamplePlugin
