+++ BaccaratTracker/Windows/ConfigWindow.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Numerics;
+using Dalamud.Interface.Windowing;
+using ImGuiNET;
+
+namespace BaccaratTracker.Windows;
+
+public class ConfigWindow : Window, IDisposable
+{
+    private Configuration configuration;
+
+    public ConfigWindow(Plugin plugin) : base(
+        "Baccarat Settings",
+        ImGuiWindowFlags.NoResize | ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoScrollbar |
+        ImGuiWindowFlags.NoScrollWithMouse)
+    {
+        this.Size = new Vector2(232, 75);
+        this.SizeCondition = ImGuiCond.Always;
+
+        this.configuration = plugin.Configuration;
+    }
+
+    public void Dispose() { }
+
+    public override void Draw() { }
+}
