+++ BaccaratTracker/Windows/MainWindow.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Numerics;
+using Dalamud.Interface.Windowing;
+using ImGuiNET;
+using ImGuiScene;
+
+namespace BaccaratTracker.Windows;
+
+public class MainWindow : Window, IDisposable
+{
+    private readonly Plugin plugin;
+
+    public MainWindow(Plugin plugin) : base(
+        "Baccarat Tracker Settings", ImGuiWindowFlags.NoScrollbar | ImGuiWindowFlags.NoScrollWithMouse)
+    {
+        this.SizeConstraints = new WindowSizeConstraints
+        {
+            MinimumSize = new Vector2(375, 330),
+            MaximumSize = new Vector2(float.MaxValue, float.MaxValue)
+        };
+        
+        this.plugin = plugin;
+    }
+    
+    public void Dispose() { }
+    
+    public override void Draw()
+    {
+        if (ImGui.Button("Show Settings"))
+        {
+            this.plugin.DrawConfigUI();
+        }
+    }
+}
